Add Config file and argument parsing support
This commit adds support for a config file to store constants. Code that
used hardcoded constants was refactored to use values from the
configuration, including adding a function to check user "permissions"
in commands.

Argument parsing support was added in a backwards compatible way. The
only positional argument is the bot token, and the config file can be
overridden with the "-c" or "--config" flag.